requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
youtube-transcript-api==0.6.2
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
joblib==1.4.2